requirements.txt
Django==2.1.15
django-authtools==1.6.0
django-model-utils==3.1.2
djangorestframework==3.10.3
coverage
flake8
